+++ src/broker/amqp.rs
@@ -78,10 +78,11 @@ impl BrokerBuilder for AMQPBrokerBuilder {
     }
 
     /// Build an `AMQPBroker`.
-    async fn build(&self) -> Result<AMQPBroker, BrokerError> {
+    async fn build(&self, connection_timeout: u32) -> Result<AMQPBroker, BrokerError> {
         let mut uri = AMQPUri::from_str(&self.config.broker_url)
             .map_err(|_| BrokerError::InvalidBrokerUrl(self.config.broker_url.clone()))?;
         uri.query.heartbeat = self.config.heartbeat;
+        uri.query.connection_timeout = Some(connection_timeout as u64);
 
         let conn = Connection::connect_uri(uri.clone(), ConnectionProperties::default()).await?;
         let consume_channel = conn.create_channel().await?;
@@ -312,14 +313,15 @@ impl Broker for AMQPBroker {
     }
 
     /// Try reconnecting in the event of some sort of connection error.
-    async fn reconnect(&self) -> Result<(), BrokerError> {
+    async fn reconnect(&self, connection_timeout: u32) -> Result<(), BrokerError> {
         let mut conn = self.conn.lock().await;
         if !conn.status().connected() {
             debug!("Attempting to reconnect to broker");
             let _consume_write_lock = self.consume_channel_write_lock.lock().await;
 
-            *conn =
-                Connection::connect_uri(self.uri.clone(), ConnectionProperties::default()).await?;
+            let mut uri = self.uri.clone();
+            uri.query.connection_timeout = Some(connection_timeout as u64);
+            *conn = Connection::connect_uri(uri, ConnectionProperties::default()).await?;
 
             let mut consume_channel = self.consume_channel.write().await;
             let mut produce_channel = self.produce_channel.lock().await;